Two new versions of pidigits for consideration
pidigits-blc.chpl was mostly me doing a review of Tom's version and
making edits as I went because that's the easiest way for me to review
code.  Some changes were an attempt to make it a purer use of Chapel
(e.g., no externs), others were to make it more elegant (by my
tastes).

Specific changes were:

* not using externs for I/O (which will cause a performance hit, but
  feels like it's keeping within the language better)
* attempted to make the looping structure more precise/less
  break-/return-oriented
* removing as many cases of specifying types/casts as often as
  possible (reducing the number from 12 to 5)
* lumped the next_term and extract_digit calls into a single function
  next_digit
* doing so permitted me to get rid of the sentinel value of -1 which
  in turn permitted me to get rid of some of the aforementioned casts
* moving variable declarations closer to their first defs/uses when
  possible
* moving the initialization of the GMP variables into the global code
  section
* adding a config const to specify whether or not to print timing and
  avoid the prediff
* adding parens to the no-arg procedures under the argument that they
  definitely seem like they compute something rather than being
  something that is subject to a space-time tradeoff
* comma-separating the standard modules used

One of the todos I noted in this version was to try and move the loop
generating the digits into an iterator to keep the top-level code as
clean as possible.  In mailing Tom I was thinking this was going to be
a bit tricky/convoluted, but it turned out to be fairly
straightforward.  The result is pidigits-iter.chpl which I think is
pretty nice.

Remaining TODOs: It would be nice to support overloaded assignment on
GMP types to avoid some of the fairly verbose initialization calls.
Related, it would be nice to support initialization functions for
external types; our qthreads work wanted this as well.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18328 3a8e244f-b0f2-452b-bcba-4c88e055c3ca