Make +/- on Rectangular domains illegal
We decided to make +/- (translate) illegal on rectangular domains
since it was inconsistent with the interpretation for irregular
domains (add/remove indices).  The translate() function on domains
should be used directly instead.  Also, added additional error
checking (no translate on irregular domains) and offsets only integral
types.

Due to unsigned integer types, tranlate must now have the ability to
take types that are not preciesly the same as their index type.  The
translation of each dimension of a domain happens in the range
translate function, so that was changed to handle the possiblity of
unsigned types.

Side note: I was intending to get rid of chpl__unTranslate (which was
documented as a an internal function), but it came in handy in the
Block and Cyclic distributions where we have to shift indices of the
each locale's portion.  It probably could have been done without
chpl__unTranslate, but with a lot of ugly conditionals.

Updated all tests that used +/- on domains and also added a few new
tests for translate.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18340 3a8e244f-b0f2-452b-bcba-4c88e055c3ca