This commit updates the specification and related files in the compiler and modules to use "Rectangular" in place of "Arithmetic".
The file DefaultArithmetic.chpl was renamed to DefaultRectangular.chpl, and
classes defined within it renamed appropriately.  Client code in
ChapelArray.chpl and the array, domain and distribution areas was also updated
to reflect this change.

Finally, the test code was updated.  There were basically two kinds of failure
that had to be corrected:  Either "arithmetic" changed to "rectangular" in a
compiler error message, or the test code itself referred directly to the
defaultRectangular classes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18343 3a8e244f-b0f2-452b-bcba-4c88e055c3ca