Rewrote sub_clean (another csh script bites the dust!)
The real reason I touched sub_clean was that I was dissatisfied with
the way it worked (removing *.tmp in a directory, inability to name a
particular test to clean, etc.).  Since it was less than 100 lines, I
just rewrote it since doing so would allow me to easily fix the
wildcard issue that Vass noticed.

sub_clean now cleans the current directory with no arguments or one
more more directories and/or files given on the command line.

Cleaning a file entails removing the following files:

         <testname>.exec.out.tmp
         <testname>.comp.out.tmp
         <testname>
         <testname>_real

In addition, files/directories listed in <testname>.cleanfiles are
also removed (NEW FEATURE).

Cleaning a directory entails cleaning all the tests *.chpl in that
directory as above, as well as removing 'core' and any files or
directories listed in CLEANFILES.

CLEANFILES and .cleanfiles may include wildcards.

start_test now calls sub_clean with a specific file name, if the file
name is listed on the command line.  Also, updated the associated
documentation regarding CLEANFILES and the new .cleanfiles.

Unrelated, but since I was in there, I added a dash-between-words
version (and a no-dash-between-words in one case) of most of the
common flags, since I was always messing that up.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18345 3a8e244f-b0f2-452b-bcba-4c88e055c3ca