Testing related to where clauses and distributions
This is a set of tests that attempts to dispatch between two
identical overloads of functions based on the distribution of
a domain or array.  

It turns out that the way that I thought this would be expressed
(in dispatchBasedOnDist.chpl) doesn't work for reasons I'm not
entirely clear on -- it could be related to the fact that .dist
on a domain is implemented as a method rather than as a value?

dispatchBasedOnDist-workaround.chpl is a workaround that dives
into the _value field (a no-no) -- ugly, but it works today.

dispatchBasedOnDist-alt.chpl is a proposed alternative syntax
if for some reason the original future can't be made to work.
The suggestion being perhaps we could have a type function
built into the domain record that one could write where
clauses on.  The main downside to this is that a user may
continue to write the original future and have it now work
for mysterious reasons.  And it's one more method to learn
about...

Ultimately, only one of these three tests should live on, which
is to say, once one of the futures is working and we're happy
with it, the other two tests can be removed.

I decided not to go further with this today, but I think the
right way to debug it would be to reproduce this case with
standalone classes/records to determine why the dispatch isn't
doing what we think would be natural.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18355 3a8e244f-b0f2-452b-bcba-4c88e055c3ca