Make the threading layer be responsible for shutting down threads at exit.
The tasking layer has been responsible for shutting down all the threads
when the tasking layer exits during program termination.  The threading
layer should actually be responsible for this instead.  So, move the
list of active threads, the maintenance code for that, and the thread
cancellation and join code into the threading layer.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18363 3a8e244f-b0f2-452b-bcba-4c88e055c3ca