Rework iterator range.these for bounded ranges, so it can be optimized again. (Correct performance regressions due to the addition of unaligned ranges.)
Add warning message for maximal range (off by default).
The behavior for a maximal range now is to execute zero times.
This is arguably OK though not strictly correct, since in most cases people will
not want to execute their loops 2^32+1 times, and the manifestation of the
problem is anything but subtle.  Nonetheless, it is a shortcoming of the
language.  The best solution is to improve the optimizer so that it can handle
short head and tail code as well as (possibly) branching out of the body of a
loop.  Those things together would make it possible to go back to the truly
correct iterator implementation.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18366 3a8e244f-b0f2-452b-bcba-4c88e055c3ca