Bug fix for copy of extern records
Prior to this commit, assignments of external record types would
sometimes copy all fields (whether or not the Chapel compiler knew
about them) and sometimes would only copy those that the Chapel
compiler knew about.  The reason for this is because some Chapel
record assignments are done via Chapel-provided overloads of
assignment and others are done via normal C record assignment (I had
completely forgotten about this, but perusing SVNLOG reminded me of
this and demonstrated Steve trying and failing to prove that the use
of C assignments was problematic in certain cases; if I'm piecing
everything together correctly, I believe that the C assignment is used
only in inferred-type initializers...).

In this commit, I fix this inconsistency by only having the compiler
implement assignment on external records for matching external record
types (i.e., an external record type R can only be assigned using
other external records of type R) and by implementing that assignment
using C-style assignment to guarantee consistency across these cases.
I've checked in a future (provided by Michael Ferguson) to demonstrate
the original issue and lock in the fix.

Another change that could be made but that I didn't make here would be
to provide a second assignment overload for external record types that
followed normal Chapel semantics (assign all its fields -- the ones
known to Chapel anyway -- by name).  Thus, for matching record types,
the C assignment would be used while the more general form would be
used for any other type.  This has the advantage of being consistent
with other Chapel record assignments while having the downsides of (a)
not assigning all fields of the record (if there are some unknown to
the Chapel compielr) and (b) being inconsistent with this
external-to-external record assignment operator.  I thought the
current approach was the more conservative thing to do for the time
being and that if people needed the other, we'd hear about it in their
complaints.  Also, arguably, a user could always implement such
assignments themselves (if we permit users to overload assignment,
which I think is the way we're trending).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18369 3a8e244f-b0f2-452b-bcba-4c88e055c3ca