Fix to implicit module naming bug + improved description
Fixing the bugs that Vass filed last week showing that we were taking
everything from the path to the first '.' in the filename in creating
an implicit module name rather than from the path to the '.chpl'
suffix.  Improved the spec wording as requested to be a bit more
precise about how the implicit module name is formed.  The main things
this required were (a) changing a strchr to an strrchr and (b) making
our identifier legalization sweep convert '.' to 'DOT' as we do for
other non-legal C identifier characters.  To make the tests work, I
also had to add a use of the --main-module flag to the second test.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18371 3a8e244f-b0f2-452b-bcba-4c88e055c3ca