Simplify sync (and single) variable handling across the threading layer interface.
We have been using pthread condition variables to implement timed waits
when sync (or single) variables are not in the desired state or the task
pool is empty.  The code for this was rather complicated and straddled
the boundary between the tasking and threading layers.  Replace it with
a spin wait with optional timeout in the tasking layer, around a yield
call to the threading layer.  This is much simpler and should improve
performance some due to smaller wakeup latencies, since the condition
variable code had to do more syscalls.  And since the threading layer
yield function is still implemented by yielding the processor, as in the
old code, the spin wait should nevertheless not waste any more cycles
than the old code did.

As a side effect of doing this, the threading layer now doesn't have to
know anything at all about sync (or single) variables or the task pool.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18375 3a8e244f-b0f2-452b-bcba-4c88e055c3ca