Committing John Lewis' final version of PTRANS
This is the version of PTRANS that John wrote to use a Block
distribution just prior to leaving Cray.  He used the Block
distribution not because it's ideal, but as a stand-in for the
not-yet-complete Block-Cyclic distribution.  This commit only has very
minor commits to his version, to remove tabs, update the 'def'
keyword, add a flag for turning off performance timings, turn down the
problem size for testing purposes, and remove an extraneous forall
loop from the 'arbitrary beta' variations.

In its current form, only the beta = 1.0 version is tested.  Ideally,
it would be nice to test beta = 0.0, 1.0 and other, but I haven't
taken that step yet.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18377 3a8e244f-b0f2-452b-bcba-4c88e055c3ca