Adding a Block-Cyclic version of John's PTRANS code
ptrans_2011-blkcyc.chpl is a variation of ptrans_2011.chpl that uses a
Block-Cyclic distribution instead of a Block distribution.
Ultimately, the Block version can be retired, but for now it is a
useful means of testing rewrites to the code when using features that
Block-Cyclic does not yet support.

Also did some cleanup in the Block version to make the diffs between
the two versions minimal.

Dialed the number of locales down to 3 simply to reduce execution time
when using a single machine (yet still test a case that isn't a
perfect square which is why I originally chose 6; 2 would also fulfill
this but seemed slightly less interesting).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18381 3a8e244f-b0f2-452b-bcba-4c88e055c3ca