Make the threading layer responsible for maintaining thread counts.
The tasking layer has been responsible for maintaining a count of the
number of threads in existence, and not creating more threads than we
are supposed to.  Move this responsibility into the threading layer.
This involves creating four new functions in the threading layer:
  threadlayer_can_start_thread()
  threadlayer_thread_destroy()
  threadlayer_get_max_threads()
  threadlayer_get_num_threads()

(Strictly speaking threadlayer_destroy() isn't necessary here, but I
added it anyway, for symmetry and to match up with the needs of tasking
layers other than the FIFO one.)

I also augmented the existing threadlayer_init() to take the maximum
number of threads per locale, in addition to its existing call stack
size argument.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18382 3a8e244f-b0f2-452b-bcba-4c88e055c3ca