* Made explicit and/or defined the following concepts:  - "function" (as the union of procedures and iterators)  - "formal", "actual", "call site"  - "operator", "method" as subsets of "function"  - "receiver" (the 'this' argument)
Note: we can change the words we use to denote these concepts, as desired.
"Function" has been the de-facto name for the procs+iters concept
in the current spec already, I let it be.

* Renamed the "Functions" chapter to "Procedures" because that's what
it focuses on. (Even though much of it applies to all "functions" as well,
as suggested by Brad.)

Rewrote the overview and added a roadmap (in Vass's new preferred format).

* Replaced 'function' with 'procedure' in the Procedures (former
Functions) chapter when appropriate.

There were a few borderline cases, e.g. "type functions" and
"param functions", and I mostly left them "functions".
(The above two are borderline because we intend to allow them, but at
present they are "unsupported", as per an "openissue" clause.)

I left "function" in "function calls" mostly as-is -- because these are
the same for procs and iters (except *how* those calls are used).

* Clarified the blank intent to correspond to what's presently implemented
in the compiler.

* Almost introduced "copy constructor" for 'in' parameter passing.
But chickened out at the last moment, especially because I had
called it "assignment operator". (Seriously, need group approval and details
- I don't think we define it anywhere in the spec.)

* Switched the order of Nested Functions and Overloading sections
moving the Nested Functions closer to other declaration-related sections
and Overloading next to Function Resolution.

* Restructured some of the Classes chapter to follow our
"define the types then define the values" guideline.

* Removed the "Class Variable Declarations" because it did not
add any information.

* Lifted "The 'this' method" and "The 'these' method" out from
the "Method Calls" sections of Classes and Records.

* Changes to the syntax rules:

 - member-access-expression is now
   a field-access-expression or method-call-expression

 - function-declaration-statement -->  procedure-declaration-statement

 - defined method-call-expression

* Added a few index entries. Some of them could be renamed
function->procedure.

* Rephrased a few pieces.

* Filled in some other gaps.

* Added some TODO notes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18389 3a8e244f-b0f2-452b-bcba-4c88e055c3ca