This should correct the compilation error on Cray and XMT C compilers. I was using the idiom
if ! param return smpn;
return smpn-else;

If an "else" does not appear explictly after the parameter-controlled
clause, then the Chapel compiler emits the follow-on code even if the
parameter expression is false.  That C code is unreachable, so some
compilers complain.  I added "else" clauses in ChapelRange.chpl, and
added a future to track this code-generation optimization opportunity.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18400 3a8e244f-b0f2-452b-bcba-4c88e055c3ca