Specify the initial function for new threads early.
Specify the function that new threads should call when we initialize the
threading layer, rather than later when we actually create threads.
Doing so allows a threading layer to create threads proactively if it
needs to, say in order to spread work across processor cores.  There is
a slight disadvantage in that all new threads must start by calling the
same function, but this is more than offset by the potential for thread
layer adaptability.  (Note that the pthreads threading layer does not
make use of this possibility for adaptive behavior yet, but it could do
so in the future.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18401 3a8e244f-b0f2-452b-bcba-4c88e055c3ca