Getting HPL foot in the door
This is a really minor commit: I'm checking in the version of HPL
that I've been working on with Block-Cyclic that doesn't use any
reindexing operators.  It currently compiles but does not yet run
due to the lack of general slicing in Block-Cyclic which is the
next thing I need to work on.  As a result I'm checking it in as
a .noexec test to avoid backsliding (e.g., it took me awhile to
realize that Tom's range change broke BlockCyclic for HPL because
the existing BlockCyclic tests didn't exercise the right code
paths.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18404 3a8e244f-b0f2-452b-bcba-4c88e055c3ca