Fixed timeout/kill and stdin issues.
timedexec: Added a secondary time out which will send SIGKILL to the
process if it does not die due to SIGTERM.  For now, it will wait for
that same timeout value passed into timedexec.  If we could tighten
this up to an absolute number if things still take too long.

sub_test: Fixed issued with redirecting stdin.  I didn't properly dupe
the code that redirected stdin when I moved back to using timedexec.
This was sometimes causing issues when taking SIGALRM in timedexec.
Also cleaned up/unified the uses of redirectin.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18406 3a8e244f-b0f2-452b-bcba-4c88e055c3ca