This commit fixes the passing of extern classes to extern procedures.
If an extern class is passed to an extern procedure, it is a reference
(C pointer) by definition so we must treat it as so.  Consequently, if
a wide reference has been inserted for that class, we must create a
temporary for the CHPL_COMM_WIDE_GET() macro (which is ultimately
dereferenced before passed to the extern procedure).  If at runtime,
the class variable is located on a different locale, an error is
issued.  Brad and I discussed these semantics and we believe them to
be what we want.

I also fixed the line numbers during the insertWideReferences pass.  I
noticed that the line numbers were not right for my error messages (I
had noticed incorrect line number before, but they never showed up in
error messages, so I ignored it).  I investigated a bit and figure out
that we use a slightly crufty scheme to get line numbers for new
expression inserted after parsing.  We set the global currentLineno
via the macro SET_LINENO() that is passed an AST node.  Any new
expressions get the global currentLineno.  It would probably be
preferable to have insertBefore/insertAfter/replace use the line
number of their object, but there are likely some assumption embedded
in the code as is, and I didn't think it was worth tackling today.

Testing:

- Added a small suite of tests that exercise extern classes including
  extern procedure, references and non-local access

- test/extern/bradc/class/extern_class.chpl regression for --no-local
  and GASNet should be fixed



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18419 3a8e244f-b0f2-452b-bcba-4c88e055c3ca