Add a thread termination function to go with the thread startup one.
This fixes an asymmetry in the tasking/threading layers.  The tasking
layer was allocating the thread-private data and assigning it to the
threads, but the threading layer was responsible for freeing it.  Move
responsibility for freeing this data into the tasking layer, so that it
handles both operations.  To do this I changed the threadlayer_init()
routine, so that it now takes pointers to two functions, one to be
called when the new thread is created, and another to be called when it
ends.  The code to free thread-private data is in the new thread-ending
function in the tasking layer, and the old code that used to do so in
the threading layer has been removed.

While here, I also renamed the function that is called when threads are
created, so that the begin and end functions have symmetric names.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18429 3a8e244f-b0f2-452b-bcba-4c88e055c3ca