Included documentation for locale.name in Locales.tex.
Updated the documentation on ranges to reflect the desired semantics
w.r.t. alignment.  Added a description of "natural alignment", to make it easy
to describe the action taken by several operators (which use the concept).

Added documentation for additional methods that have been
added to support the expanded definition of the count operator (taking negative
counts).  These include: hasFirst(), hasLast() and ident().
I also added documentation for other methods that had been left out on
previous passes: alignHigh(), alignLow(), ==, boundsCheck(), etc.

Added the "big table" as a Quick Reference section at the end of the chapter.  I
wrote the table as if "absolute alignment" has already been implemented.

Since it turned out that most of our discussion about alignment was really about
the semantics of the by and # operators, it turns out the actual value stored in
the alignment field is mostly hidden from the user.  Whether it is "natural" or
"absolute" is only exposed in the align and writeThis routines.  In the latter,
it is even hidden in most cases, since the semantics described in the spec say
that the "align" suffix is suppressed if the alignment is natural.  

So, what I'm getting at is that originally the "big table" was supposed to
contain representations for the resulting range in terms of both absolute and
natural alignment.  But after I went back and implemented the ultimately desired
semantics, the distinction becomes far less interesting: It is just an
implementation detail.

I still think that it is worth the trouble of converting the internal
representation to absolute.  It will simplify some of the code and make the
whole module more maintainable.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca