Changed the leader iterator of the global array descriptor classes to simply redirect to the leader of the global domain descriptor, for these domain maps: - DefaultRectangular - Block - Cyclic
Motivation: reduce code duplication.

I verified that the matching domain and array leaders prior to this change
had been essentially identical. Each array leader included references to
the corresponding domain throughout its code; now there is a single
reference to the domain (for the redirect).

This will change the observed behavior when debugging output is enabled:
before each leader indicated whether it was for the domain or the array,
now the domain leader output will show even when the array leader is invoked.
This should not affect end users, however.

If performance changes, we should reconsider this.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18432 3a8e244f-b0f2-452b-bcba-4c88e055c3ca