Minor improvements to Block-Cyclic
This commit makes some minor improvements to Block-Cyclic that I did
"for fun" while on vacation instead of looking at email.

The main advance here is to add some initial support for slicing
Block-Cyclic distributed arrays that supports some cases that I've
tested and also seems to permit the version of HPL that I've been
focusing on to compile and run using a Block-Cyclic distribution.
That said, upon returning to town and trying it on multiple locales,
I'm finding that something about my approach doesn't work in that case
-- I think the failure has to do with cases where a locale doesn't own
any indices for the given slice but still need to debug.  This felt
like a sufficient positive step forward (and won't make anything fail
that didn't before) that I thought it was worth committing.

The main approach that I took here was one that I've envisioned for
Block-Cyclic since I stopped working on it ages ago, but have never
had the chance to implement.  Essentially, all Block-Cyclic arrays
store two domains -- one indicating how they are allocated in memory
and a second indicating the logical indices that they describe.  So
upon slicing a Block-Cyclic domain, the allocated domain is inherited
from the original array (since the allocation remains the same) while
the logical domain is whatever was used to do the slicing.  This
approach seems to be more necessary with Block-Cyclic than for Block
and Cyclic for reasons that I haven't been able to fully describe well
yet -- but I think it has to do with the fact that regular subsets of
them are not trivially describable using the same parameters that were
used to describe the original array.  This is a pattern I think we'll
see more and more going forward (e.g., I expect that sparse slicing
will need to take a similar approach).

Next step: Figure out the multi-locale case and then convert these
tests to use multiple locales.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18442 3a8e244f-b0f2-452b-bcba-4c88e055c3ca