This checkin just clears the decks, so I can rework Ranges once again to take into account the recent discussions regarding partially bounded ranges, and how to figure out alignment.
Taking into account the feedback from the users mailing list, it appears
possible to accept partially-bounded ranges with "wrong-way" strides, and not
have to determine alignment arbitrarily.  Stay tuned.

Standardized on "sequence" rather than "index set" as what a range describes.
Standardized on "index/indices" as what a range contains, rather than "value", "member" and other terms.

Added a description of the count operator according to Vass's suggestion of
retaining one of the bounds.  I like the idea in the abstract, but ran into
regressions only w.r.t. the tests in studies/amr.  So I'll have to approach that
one more incrementally.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18466 3a8e244f-b0f2-452b-bcba-4c88e055c3ca