Improved reading of timeout values.
Properly open the file and read all the contents.  I had originally
just copied the old sub_test and used 'cat' (I'm so embarrassed!)

Timeout files can now contain whole line comments starting with #.
The first line to not start with a # must be the (integer) timeout
value (with nothing else on the line).  You can put whatever you want
in the rest of the file.

Fixed the Fatal() subroutine/procedure/function/whatever to print
error messages in the "standard" format so that they will be reported
by the summary.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18479 3a8e244f-b0f2-452b-bcba-4c88e055c3ca