Bug fix for parallel iteration over strided Block-Cyclic Domains
This commit fixes parallel iteration over strided Block-Cyclic
domains.  Previously, I forgot to take the stride of the domain into
account which resulted in having the follower iterate too much by a
factor of stride.

Added a test that locks in this behavior -- loopBlkCyc.chpl.

With this test, HPL now runs on multiple locales, so I added a
numlocales file to my draft of HPL.  I also dialed down the problem
size because it takes a fair amount of time to run currently.  More
still needs to be done to tune HPL for locality and tune Block-Cyclic,
but this is a great first step.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18485 3a8e244f-b0f2-452b-bcba-4c88e055c3ca