Adding initial support for qthreads
This commit commits version 1.5.1 of qthreads which is the first to
support Chapel out-of-the-box.  It also commits the supporting
Makefiles that will copy the "shim" code (technically Chapel runtime
code that lives in the qthreads distribution) out of the qthreads
release and into our runtime directory structure.

Note that qthreads currently only supports single-locale/non-GASNet
execution.  We are currently working on adding support for
multi-locale qthread execution.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18487 3a8e244f-b0f2-452b-bcba-4c88e055c3ca