Updates to the Cyclic distribution for recent changes to Ranges.
Fixed 3 places where ranges are used to set up other ranges.

In 2 of the cases (getChunk() and the leader iterator), the
intermediate ranges sometimes resulted in non-zero aligment.  Though
they represented the "correct" values, the final range was assumed to
have an alignment of 0.  In the interest of minimizing the changes
before the upcoming release, this was fixed by removing the alignment,
though in the future, we may want to review these 2 bits of code.

In the other case (LocCyclic constructor), intermediate ranges were
used to set up a range that spanned the entire index with the correct
stride (i.e., explicitly bounded).  This case simply required use of
the correct value for the bounds (last) which had changed due to the
change in meaning of of low and high.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18490 3a8e244f-b0f2-452b-bcba-4c88e055c3ca