Updating how STATUS is maintained and structure of top-level Makefile
As discussed on email these past few days, updating STATUS so that it
isn't stored in SVN, but rather recomputed on demand.  To this end,
deleted STATUS from SVN, added a rule to the gen_release script to
create it, and added it to the svn:ignore list for the root directory.

At the same time I did a Makefile refactoring that's been on my TODO
list for a long time by moving all developer-oriented calls into a
sub-Makefile called Makefile.devel.  This currently includes the rules
to make the man page, to make STATUS, and to make the spec tests.
This file is not included in the release and so its rules will only be
available to those working from the SVN tree ("developers") by
default.

It occurs to me as I type this that it might be more useful to not
have STATUS get rebuilt by default and to not add it to the svn:ignore
list so that (a) you might not have a local copy of STATUS by default
and (b) if you accidentally update a local copy of STATUS, you'll see
that it's marked as ? rather than M in an svn status.  The fear being
that you may make local edits to STATUS and then forget and never
notice that they need to be propagated to STATUS.devel.
Comments/preferences?  (Note that most of the work in this
5-minute-change (TM) is still useful in either case).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18492 3a8e244f-b0f2-452b-bcba-4c88e055c3ca