Adding Nanox to SVN
This adds a snapshot of Nanox from their git sources to our SVN log to
lock in a working version for the upcoming release.  This snapshot was
"blessed" by Alex Duran at BSC for use in our release (their 0.5
release wasn't in sync with our code base and they weren't ready for
their 0.6 release yet).  See nanox/README for details.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18528 3a8e244f-b0f2-452b-bcba-4c88e055c3ca