paratest.server now filters the output of paratest.client to show only those error messages that have been explicitly produced by paratest.client via calling its fatal() function.
Rationale: this is better than showing everything that has 'error'
in it (e.g. numerous compiler errors). The downside is that
paratest.server may hide some errors that would be useful to see,
but I do not know of any at present.

Ideally we would like paratest.server to show ssh-related messages
(e.g. "The authenticity of host '...' can't be established. [...]
Are you sure you want to continue connecting (yes/no)?" and
"user@host's password:"). The issue here is those prompts
do not end with a newline so 'grep' would not show them immediately.
But let it be for now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18538 3a8e244f-b0f2-452b-bcba-4c88e055c3ca