* Renamed (for a more adequate name):   chpl_privateInstance -> chpl_getPrivatizedCopy
* Replaced all uses of __primitive("chpl_getPrivatizedClass")
with calls to chpl_getPrivatizedCopy()
(in modules and tests).

BTW this allowed us to eliminate this idiom:
  var distpid = dist.pid;
  var thisdist = dist;
  var privdist = __primitive("chpl_getPrivatizedClass", thisdist, distpid);
because (seemingly) the values get widened/unwidened (whichever is needed
here) when passing as arguments.

* Removed the implementation of chpl_privateInstance/chpl_getPrivatizedCopy
that accepts the original object as the value
(therefore all uses of chpl_getPrivatizedCopy are required
to pass a type as the first argument).

Rationale: this matches what the primitive currently does.
There seem to be no plans to change that, so we force
all future uses to stick to this convention.
Moreover, this makes it clear at the callsite that
the actual value is not used/referenced.

* Updated README.dsi correspondingly.
Also replaced 'instance' -> 'object', for consistency.

* Also: removed the requirement for GlobalArray.idxType from the interface
(by replacing its uses in ChapelArray.chpl with dom.idxType).
Correspondingly, removed its implementation from ReplicatedDist
and from README.dsi.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca